Align behavior of menu items with toolbar items
In previous releases, the "File->New" menu opened a new connection dialog and opened a new tab for it.
After the introduction of the Toolbar, the behavior changed and it just opened a new tab.
This change makes it open a new connection dialog (but not open a new tab for it). Just like the "Toolbar->New connection" button does.

"Toolbar->New file" actually opened a new connection. Now it opens a new tab as there already is a button for a new connection.

There are also some unifications in button/menu descriptions.