Added version folders
Brought back the previous API documentation in a 2018.4 version folder and added a 2019.2 folder for the updated documentation.  Placed the original documentation in the schemas folder to preserve the original URLs.  Added a very simple index.html page in the root to allow the selection of the version to view.