Updated SwqlStudio.csproj to use Serilog for logging
Updated SwqlStudio.csproj to eliminate the SolarWinds.Logging dependency (and a transitive dependency on log4net.  It is configured instead to use Microsoft.Extensions.Logging and Serilog for logging.  The log file path is no longer configurable via the App.config file in this approach.  Updated the installer to add the new dependencies, including transitive dependencies.  The SolarWinds.Logging and log4net assemblies are still deployed since SolarWinds.InformationService.Contract still depends on them.  This is not ready for merging.